feat: generalize open
