+++ src/lib/buildKey.ts
@@ -1,0 +1,6 @@
+const buildKey = (appName: string, title: string) => {
+  return `${appName}/${title}`
+}
+
+
+export default buildKey;
